impr(inflation): add grpc query + tests

+++ x/inflation/keeper/epoch_info.go
@@ -39,19 +39,19 @@ func (k Keeper) SetEpochsPerPeriod(ctx sdk.Context, epochsPerPeriod int64) {
 	store.Set(types.KeyPrefixEpochsPerPeriod, sdk.Uint64ToBigEndian(uint64(epochsPerPeriod)))
 }
 
-// GetSkippedEpochs gets the epochs per period
-func (k Keeper) GetSkippedEpochs(ctx sdk.Context) int64 {
+// GetSkippedEpochs gets the number of skipped epochs
+func (k Keeper) GetSkippedEpochs(ctx sdk.Context) uint64 {
 	store := ctx.KVStore(k.storeKey)
 	bz := store.Get(types.KeyPrefixSkippedEpochs)
 	if len(bz) == 0 {
 		return 0
 	}
 
-	return int64(sdk.BigEndianToUint64(bz))
+	return sdk.BigEndianToUint64(bz)
 }
 
-// SetSkippedEpochs stores the epochs per period
-func (k Keeper) SetSkippedEpochs(ctx sdk.Context, skippedEpochs int64) {
+// SetSkippedEpochs stores the number of skipped epochs
+func (k Keeper) SetSkippedEpochs(ctx sdk.Context, skippedEpochs uint64) {
 	store := ctx.KVStore(k.storeKey)
-	store.Set(types.KeyPrefixSkippedEpochs, sdk.Uint64ToBigEndian(uint64(skippedEpochs)))
+	store.Set(types.KeyPrefixSkippedEpochs, sdk.Uint64ToBigEndian(skippedEpochs))
 }
